refactor(agent_backends): extract gateway submodule
Move the per-backend TCP/HTTP gateway and the BackendGateway dispatcher
out of the facade and into gateway.rs:

- BackendGateway (pub re-exported through mod.rs so AppState's
  `pub backend_gateway: BackendGateway` import path stays
  `crate::commands::agent_backends::BackendGateway`)
- GatewayServer, generate_gateway_token
- run_gateway, handle_gateway_connection
- gateway_route_requires_auth, gateway_auth_matches
- route_path, find_header_end
- write_json_response, write_empty_response
- write_json_or_sse_response, anthropic_sse_body

gateway_translate's `write_anthropic_error_response` now delegates to
`super::gateway::write_json_response`, which keeps the single
HTTP-writer implementation in gateway.rs while the translation layer
that needs to emit the Anthropic error envelope still owns the
status-to-error-type mapping.

Tests in mod.rs gain a `use super::gateway::{anthropic_sse_body,
gateway_auth_matches, gateway_route_requires_auth, route_path};`
import; they will follow into gateway.rs alongside their primaries
in the facade-trim commit.

cargo fmt --check, cargo clippy on the three CI crates with
-Dwarnings, and the agent_backends tests (111 passed) all pass.
No public-surface change. AppState's BackendGateway field, the
disabled-twin BackendGateway, and the 9 #[tauri::command] entry
points are all unchanged.
